Allow VPN lockdown UID ranges to stack properly
When updating lockdown UID ranges, do not remove a UID from lockdown
if it is still present in any of the previously-supplied ranges that
have yet to be removed. This allows supplied lockdown ranges to stack
properly, providing an assurance that a UID will remain subject to
lockdown until all of the ranges including it have been removed.

Change-Id: Ia95724cd19040f83cea2c169a2585ab5dbdddbac