Moving NPMS related tests to cts
Developing these tests from within this project causes a lot of
maintenance toil because they are not applicable to the code in the
module. Moving them to be in the cts repo where they're on the same
development cycle as the code they test.

The tests are being moved from tests/cts/hostside-network-policy to
hostsidetests/network-policy to be consistent with the destination
repository's naming conventions.

Test: atest CtsHostsideNetworkPolicyTests

Bug: 322115994
Change-Id: I53bb4a25ef8f3229e8cf5024c59f4f1aa8c63cce