Fix setEthernetEnabled API to properly affect administrative state
Previously, setEthernetEnabled(false) untracks all interfaces and
ignores future RTM_NEWLINK callbacks, which means that the ethernet
module loses all knowledge of existing interfaces.

This change now properly toggles IFF_UP on all interfaces including
those in server mode when setEthernetEnabled API is called, and returns
on error on callback when trying to enable an interface when ethernet is
disabled.

Test: atest EthernetManagerTest
Bug: 316268558
Bug: 225315248
Change-Id: Ic7785e7e0c87c1437627661401e66201dd60802a